+++ Inc/bma280.h
@@ -134,12 +134,12 @@
 class BMA280
 {
   public:
-  BMA280(GPIO_TypeDef* SPICSTypeDef, uint16_t SPICSPin,SPI_HandleTypeDef* bmaspi,uint16_t BaseID);
+  BMA280(GPIO_TypeDef* SPICSTypeDef, uint16_t SPICSPin,SPI_HandleTypeDef* bmaspi,uint32_t BaseID);
   void init(uint8_t aRes,uint8_t BW, uint8_t power_Mode, uint8_t sleep_dur);
   void reset();
   int getData(DataMessage * Message,uint64_t RawTimeStamp,uint32_t CaptureCount);
   int getDescription(DescriptionMessage * Message,DescriptionMessage_DESCRIPTION_TYPE DESCRIPTION_TYPE);
-  int setBaseID(uint16_t BaseID);
+  int setBaseID(uint32_t BaseID);
   uint8_t getChipID();
   uint8_t getTapType();
   uint8_t getTapStatus();
@@ -156,7 +156,7 @@ class BMA280
   uint16_t _SPICSPin;
   SPI_HandleTypeDef* _bmaspi;
   uint32_t _ID;
-  uint16_t _BaseID;
+  uint32_t _BaseID;
   uint16_t _SetingsID;
   uint8_t _aRes;
 };
